Extract componentes from FixedNavbar
* Extract Link component to separate not have two components
  in the same file.

* Rename FixedNavbar to Navbar to maintain the nave convention used.
